back_end/gcloud-functions/requirements.txt
kubernetes
google-cloud
google-cloud-container
google-cloud-storage
PyYAML
cloudstorage
flask
requests